修改了entity中bean的id类型为Long

+++ src/java/com/viv/dao/UserProjectOperation.java
@@ -18,6 +18,6 @@ public interface UserProjectOperation {
     /*根据userId和project确定，选择一个userProject记录*/
     public List<User_project> selectByUseridProjectid(Map map);
     /*根据UserId返回记录数量*/
-    public int countByUserId(int user_id);
+    public int countByUserId(Long user_id);
 
 }
